13:30

+++ PYTHON/onlineshopping3/app.py
@@ -4,7 +4,7 @@ from flask_login import LoginManager, UserMixin, login_user, login_required, log
 from flask_wtf import FlaskForm
 from wtforms import StringField, PasswordField, SubmitField
 from wtforms.validators import DataRequired
-from wtforms.fields.simple import FloatField
+from wtforms.fields.simple import DecimalField
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'
@@ -54,7 +54,7 @@ class SupplierLoginForm(FlaskForm):
 class ProductForm(FlaskForm):
     name = StringField('Product Name', validators=[DataRequired()])
     description = StringField('Product Description')
-    price = FloatField('Product Price', validators=[DataRequired()])
+    price = DecimalField('Product Price', validators=[DataRequired()])
     submit = SubmitField('Add Product')
